[COMMON] move ClusterInfo#Optimized to OptimizedClusterInfo (#1566)
